Added comments.
The algorithm used here, solves sudoku by recording all possible numbers for a cell that satisfy the condition for row, column and 3x3 block (sub-grid). It assigns one possibility to the empty cell and makes a recursive call. If there is no possibility then backtrack and change previous empty cell with another of its possibility if any. This algorithm may take a lot of time to compute some sudoku puzzles, especially those with less number of given clues like 17-25.